Use relative path for systemd-run
The bash executable is not always located at /usr/bin/bash (such as on
Ubuntu 18.04 systems). systemd-run only requires an absolute path when
running a .service unit, from the docs:

https://www.freedesktop.org/software/systemd/man/systemd-run.html
> If a command is run as service unit, the first argument needs to be
an absolute program path.

Because the tomb.bash is already using env in the shebang, bash is
likely to be on the PATH.